Added sdr.py for parsing packets out of FM streams
This was actually written before the config and parse stuff, which
explains the duplicate whitening function and lack of configuration use.
It was also written as I went, and my experience with RF is minimal at
best, which explains the code quality... Basically, this was the first
version that worked at acceptable speed.